android: Move preferences class out of settings
