+++ android/app/src/main/java/org/musicpd/Receiver.java
@@ -13,12 +13,12 @@ public class Receiver extends BroadcastReceiver {
 	public void onReceive(Context context, Intent intent) {
 		Log.d("Receiver", "onReceive: " + intent);
 		if (intent.getAction() == "android.intent.action.BOOT_COMPLETED") {
-			if (Settings.Preferences.getBoolean(context,
-							    Settings.Preferences.KEY_RUN_ON_BOOT,
+			if (Preferences.getBoolean(context,
+							    Preferences.KEY_RUN_ON_BOOT,
 							    false)) {
 				final boolean wakelock =
-					Settings.Preferences.getBoolean(context,
-									Settings.Preferences.KEY_WAKELOCK, false);
+					Preferences.getBoolean(context,
+									Preferences.KEY_WAKELOCK, false);
 				Main.start(context, wakelock);
 			}
 		}
